requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
pydantic==2.9.0
pydantic-settings==2.5.0
pyyaml==6.0.2
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.18
httpx==0.27.2
pillow==10.4.0
moviepy==1.0.3
aiofiles==24.1.0
azure-identity>=1.19.0

